refactor: getPlainText file rename & typing
