+++ src/app/(frontend)/trip-reports/_components/TripReportCard.tsx
@@ -2,7 +2,7 @@ import React, { Suspense } from 'react'
 import Image from 'next/image'
 import Link from 'next/link'
 
-import { getPlainText } from '@/lib/utils/getPlainText'
+import { getPlainText } from '@/lib/utils/get-plain-text'
 import { TripReport } from '@/payload-types'
 import type { Exec } from '@/payload-types'
